more computation refactoring
